webdriverio: finalizing the actions class
