+++ packages/webdriverio/src/commands/browser/actions.ts
@@ -1,5 +1,5 @@
-import { Actions } from 'src/classes/actions'
+import { Actions } from '../../scripts/actions'
 
-export default function actions (this: WebdriverIO.Browser) {
+export default function actions(this: WebdriverIO.Browser) {
     return new Actions(this)
 }
